+++ common/src/states/utils.rs
@@ -187,6 +187,7 @@ impl Body {
                 arthropod::Species::Emberfly => 75.0,
             },
             Body::Crustacean(_) => 80.0,
+            Body::Plugin(body) => body.base_accel(),
         }
     }
 
@@ -260,6 +261,7 @@ impl Body {
             Body::Ship(_) => 6.0 / self.dimensions().y,
             Body::Arthropod(_) => 3.5,
             Body::Crustacean(_) => 3.5,
+            Body::Plugin(body) => body.base_ori_rate(),
         }
     }
 
@@ -305,6 +307,7 @@ impl Body {
                 Body::QuadrupedSmall(_) => 1500.0 * self.mass().0,
                 Body::Arthropod(_) => 500.0 * self.mass().0,
                 Body::Crustacean(_) => 400.0 * self.mass().0,
+                Body::Plugin(body) => body.swim_thrust()?,
             } * front_profile,
         )
     }